Redo how we make room for header
Login, register, and forgot password don't have header from the
main app view, and in these cases body tag shouldn't have top
padding. Since adding a class to the body would be too much of a
hassle from react point of view, we're adding a div for routes
that need the padding.